#1 Refactor contents of __init__.py to sep. files
